Fix for WFCORE-2174. Move AwaiterModelControllerClient out of .impl package
