+++ cli/src/main/java/org/jboss/as/cli/AwaiterModelControllerClient.java
@@ -19,13 +19,14 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.jboss.as.cli.impl;
+package org.jboss.as.cli;
 
 import java.io.IOException;
-import org.jboss.as.cli.CommandLineException;
 import org.jboss.dmr.ModelNode;
 
 /**
+ * ModelControllerClient set on CommandContext must implement this interface.
+ * That is required for reload and shutdown commands to be used.
  *
  * @author jdenise@redhat.com
  */
